Move generate_pre_keys into pre_keys.rs
